requirements.txt
appdirs==1.4.3
Babel==2.4.0
Django==2.1.2
django-countries==5.0
django-dotenv==1.4.2
django-phonenumber-field==1.2.0
djangorestframework==3.7.7
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
packaging==16.8
phonenumberslite==8.4.0
promise==2.1
psycopg2==2.7.3.2
pyparsing==2.2.0
pytimeparse==1.1.7
pytz==2017.2
Rx==1.6.0
singledispatch==3.4.0.3
six==1.10.0
typing==3.6.2
